Move EPUBMediaType to separate file.
